refactor(codeblock): refactor
